Fix type conversion and Swift 6 concurrency errors
Fix compilation errors for Swift 6 strict concurrency mode:

Type Conversion Fixes:
• SCNMatrix4 init: Convert Float to CGFloat explicitly
• All 9 matrix elements now wrapped in CGFloat()

Swift 6 Concurrency Fixes:
• Add Sendable conformance to Mesh
• Add Sendable conformance to Mesh.Triangle
• Add Sendable conformance to ChokeStandard
• Add Sendable conformance to ChokeTestResult

Why Sendable is needed:
- Structs passed between actors must be Sendable
- Actor (ChokeTestAnalyzer) uses these structs across isolation boundaries
- Swift 6 enforces strict concurrency checking
- All our structs are value types → safe to send

This fixes errors:
• Cannot convert Float to CGFloat (9 instances)
• Main actor-isolated instance method cannot be called
• Main actor-isolated property cannot be referenced

+++ ChokeTestApp/ChokeTestApp/Views/MeshPreviewView.swift
@@ -189,10 +189,10 @@ struct SceneKitView: NSViewRepresentable {
 extension SCNMatrix4 {
     init(_ m: simd_float3x3) {
         self.init(
-            m11: m[0][0], m12: m[0][1], m13: m[0][2], m14: 0,
-            m21: m[1][0], m22: m[1][1], m23: m[1][2], m24: 0,
-            m31: m[2][0], m32: m[2][1], m33: m[2][2], m34: 0,
-            m41: 0,       m42: 0,       m43: 0,       m44: 1
+            m11: CGFloat(m[0][0]), m12: CGFloat(m[0][1]), m13: CGFloat(m[0][2]), m14: 0,
+            m21: CGFloat(m[1][0]), m22: CGFloat(m[1][1]), m23: CGFloat(m[1][2]), m24: 0,
+            m31: CGFloat(m[2][0]), m32: CGFloat(m[2][1]), m33: CGFloat(m[2][2]), m34: 0,
+            m41: 0,                m42: 0,                m43: 0,                m44: 1
         )
     }
 }
